backend/requirements.txt
python-dotenv==1.2.1
fastapi==0.121.3
uvicorn==0.38.0
pydantic-settings==2.12.0
pydantic-ai==1.20.0
openai==2.8.1
sambanova==1.2.0
python-multipart==0.0.20
Pillow==12.0.0